+++ dataloaders/__init__.py
@@ -1,5 +1,5 @@
 from dataloaders.simple_text_dataset import SimpleTextDataset
-from dataloaders.video_dataset import VideoDataset, custom_collate_fn
+from dataloaders.video_clip_dataset import VideoDataset, custom_collate_fn
 from dataloaders.stats_dataset import StatsDataset, stats_collate_fn
 
 __all__ = [
